+++ spec/langchain/llm/cohere_spec.rb
@@ -2,7 +2,7 @@
 
 require "cohere"
 
-RSpec.describe LLM::Cohere do
+RSpec.describe Langchain::LLM::Cohere do
   let(:subject) { described_class.new(api_key: "123") }
 
   describe "#embed" do
